[chore] Move tests related to `component.Type`
